# lib: copy arguments object instead of leaking it