--- a/lib/_http_client.js
+++ b/lib/_http_client.js
@@ -608,10 +608,18 @@ ClientRequest.prototype.setTimeout = function(msecs, callback) {
 };
 
 ClientRequest.prototype.setNoDelay = function() {
-  this._deferToConnect('setNoDelay', arguments);
+  const argsLen = arguments.length;
+  const args = new Array(argsLen);
+  for (var i = 0; i < argsLen; i++)
+    args[i] = arguments[i];
+  this._deferToConnect('setNoDelay', args);
 };
 ClientRequest.prototype.setSocketKeepAlive = function() {
-  this._deferToConnect('setKeepAlive', arguments);
+  const argsLen = arguments.length;
+  const args = new Array(argsLen);
+  for (var i = 0; i < argsLen; i++)
+    args[i] = arguments[i];
+  this._deferToConnect('setKeepAlive', args);
 };
 
 ClientRequest.prototype.clearTimeout = function(cb) {