--- a/lib/_tls_wrap.js
+++ b/lib/_tls_wrap.js
@@ -960,7 +960,11 @@ function normalizeConnectArgs(listArgs) {
 }
 
 exports.connect = function(/* [port, host], options, cb */) {
-  var args = normalizeConnectArgs(arguments);
+  const argsLen = arguments.length;
+  var args = new Array(argsLen);
+  for (var i = 0; i < argsLen; i++)
+    args[i] = arguments[i];
+  args = normalizeConnectArgs(args);
   var options = args[0];
   var cb = args[1];