--- a/lib/console.js
+++ b/lib/console.js
@@ -85,7 +85,10 @@ Console.prototype.trace = function trace() {
 
 Console.prototype.assert = function(expression) {
   if (!expression) {
-    var arr = Array.prototype.slice.call(arguments, 1);
+    const argsLen = arguments.length || 1;
+    const arr = new Array(argsLen - 1);
+    for (var i = 1; i < argsLen; i++)
+      arr[i - 1] = arguments[i];
     require('assert').ok(false, util.format.apply(null, arr));
   }
 };