--- a/lib/net.js
+++ b/lib/net.js
@@ -59,7 +59,11 @@ exports.createServer = function(options, connectionListener) {
 // connect(path, [cb]);
 //
 exports.connect = exports.createConnection = function() {
-  var args = normalizeConnectArgs(arguments);
+  const argsLen = arguments.length;
+  var args = new Array(argsLen);
+  for (var i = 0; i < argsLen; i++)
+    args[i] = arguments[i];
+  args = normalizeConnectArgs(args);
   debug('createConnection', args);
   var s = new Socket(args[0]);
   return Socket.prototype.connect.apply(s, args);
@@ -858,7 +862,11 @@ Socket.prototype.connect = function(options, cb) {
     // Old API:
     // connect(port, [host], [cb])
     // connect(path, [cb]);
-    var args = normalizeConnectArgs(arguments);
+    const argsLen = arguments.length;
+    var args = new Array(argsLen);
+    for (var i = 0; i < argsLen; i++)
+      args[i] = arguments[i];
+    args = normalizeConnectArgs(args);
     return Socket.prototype.connect.apply(this, args);
   }