Instead of leaking the arguments object by passing it as an argument to a function, copy it's contents to a new array, then pass the array. This allows V8 to optimize the function that contains this code, improving performance.

### Benchmarks

Using [benchmark.js](http://benchmarkjs.com/)

#### ClientRequest.setNoDelay / ClientRequest.setSocketKeepAlive

before: 15,286,485 ops/sec
after: 20,160,118 ops/sec
~33% faster

#### net.connect / tls.connect

before: 315,000 ops/sec
after: 345,000 ops/sec
~10% faster

code:
```js
net.connect({port: 80, host: 'google.com'}, () => {});
// same for tls
```

#### assert.throws

before: ~~98,403 ops/sec~~ 89,592 ops/sec
after: ~~145,609 ops/sec~~ 146,464 ops/sec
~~~48% faster~~ ~63% faster

code:
```js
function throws() {
  throw Error('error message');
}
// in test:
assert.throws(throws);
```

#### assert.doesNotThrow

before: ~~495,686 ops/sec~~ 424,867 ops/sec
after: ~~7,807,355 ops/sec~~ 14,513,552 ops/sec
~~~1475% faster~~ ~3316% faster

code:
```js
function doesNotThrow() {
  return 1 + 3;
}
// in test:
assert.doesNotThrow(doesNotThrow);
```

#### console.assert

before: 2,498,701 ops/sec
after: 4,163,156 ops/sec
~67% faster

code:
```js
console.assert(true);
```