@reqshark 2015-12-20
> Instead of leaking the arguments object by passing it as an argument to a function, copy it's contents to a new array, then pass the array

intuitively, asking the system to make extra copies does not seem to be the sort of programming that results in such speedup. of course the v8 compiler performs code transforms on our behalf as well as other robust forms of alchemy.

I heard the arrow functions are an improvement on regular `bind()` latency, you mentioned doing:

```js
net.connect({port: 80, host: 'google.com'}, () => {});
```

can you point me to the benchmark code, I would like to check it out?