@ChALkeR 2015-12-20
@woollybogger You are targeting the `master` branch, which already [has v8 4.7](https://github.com/nodejs/node/pull/4106).
v8 4.7 [ships](http://v8project.blogspot.com/2015/10/v8-release-47.html) [rest parameters](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Functions/rest_parameters) enabled by default.

Perhaps using those would be better/faster than copying `arguments`?