@TheAlphaNerd 2015-12-20
is this something we want to backport? If so using rest parameters would make that more difficult