@ChALkeR 2015-12-20
@TheAlphaNerd Is there a reason _not_ to use rest parameters for the sake of backporting?