@ChALkeR 2015-12-20
Rest parameters should turn this:
```javascript
Console.prototype.assert = function(expression) {
  if (!expression) {
    var arr = new Array(argsLength - 1);
    for (var i = 1; i < argsLength; i++) {
      arr[i - 1] = arguments[i];
    }
    require('assert').ok(false, util.format.apply(this, arr));
  }
}
```
into this:
```javascript
Console.prototype.assert = function(expression, ...arr) {
  if (!expression) {
    require('assert').ok(false, util.format.apply(this, arr));
  }
}
```