@mscdex 2015-12-20
I thought rest parameters were still slow in v8?