@ChALkeR 2015-12-20
@mscdex In a single micro-benchmark, rest parameters are indeed about 2% slower than copying `arguments` in Node.js v5.3.0 (v8 4.6). I did not try different cases, though. But they were not shipped by default then, and I did not check `master` branch (v8 4.7) yet.