@ChALkeR 2015-12-20
@mscdex Tested `master` branch, v8 4.7, and a few different cases (no arguments, several arguments, various types). It appears that you are correct and that rest parameters are indeed slow atm.

@woollybogger Disregard my previous comments about rest parameters for now, perhaps those will be good to use in some future v8 version.