@jasnell 2015-12-24
Marking as a watch for v4.x but we'll need to see how this ends up before landing it there.