@jasnell 2016-01-15
CI: https://ci.nodejs.org/job/node-test-pull-request/1245/