@jasnell 2016-01-15
@woollybogger ... it's kind of like playing the powerball lottery, if you roll the dice and get all green, you win big! ... unfortunately there's some current wonkiness with the windows machines due to a number of issues (a bad commit that's being looked at + unrelated build machine failures).

given that LTS is mostly green and the failures look unrelated, LGTM!