@woollybogger 2016-02-19
Bump. Resolved merge conflicts.