@woollybogger 2016-02-24
I completely rewrote my changes for [assert.js](https://github.com/nodejs/node/pull/4361/files#diff-2) and updated the benchmark results.