@woollybogger 2016-03-02
Updated to use `const` where possible