@rvagg 2016-03-09
I'm fine with this although I'm not a huge fan of manually copying `arguments`, there's plenty of precedent in core for the manual approach already so it's not a big deal but how much more costly is a `Array.prototype.slice.call()` in this situation?