@mscdex 2016-03-09
@rvagg Using `Array.prototype.slice.call()` will cause the same performance problem (passing `arguments` to any function other than a `fn.apply()`).

CI again: https://ci.nodejs.org/job/node-test-pull-request/1887/