@williamkapke 2016-03-09
Can we consider using an inline-able helper to DRYup the code?
```javascript
function toArray() {
  var args = new Array(arguments.length);
  for (var i = 0; i < arguments.length; i++)
    args[i] = arguments[i];
  return args;
}

function neeto() {
  var args = toArray.apply(this, arguments);//doesn't deoptimize
  //use 'em
  console.log(args);
}
```