@mscdex 2016-03-09
@williamkapke `.apply()` just makes things slower. Not worth it IMHO.