@williamkapke 2016-03-09
@mscdex Normally, yes, apply is slower- however with it inlined it doesn't matter.... right? (I'm happy to accept I'm completely wrong on this)