@mscdex 2016-03-09
@williamkapke It wouldn't get inlined. Even if it would, I probably wouldn't want to depend on it.