@williamkapke 2016-03-09
@mscdex Gah- just ran with `--trace-inlining` and indeed it didn't inline now. I suspect this might have been something I observed when I was checking out what I could get TurboFan to do. Please disregard my suggestion for now since I can't remember how I got it to inline :D